requirements.txt
asgiref==3.8.1
Django==4.1.13
djangorestframework==3.13.1
drf-yasg==1.21.7
flake8==7.0.0
flake8-isort==6.1.1
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
packaging==24.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
